requirements.txt.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pymongo==4.5.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-dotenv==1.0.0
beanie